Refactor for packaging.
Add pyproject.toml
Add setup.cfg

No license listed as have no decided license yet.
