+++ setup.cfg
@@ -1,0 +1,27 @@
+[metadata]
+name = parseplot
+version = 0.0.1
+author = David Antonini
+author_email = toonarmycaptain@hotmail.com
+description = Maths expression parser-plotter.
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/toonarmycaptain/parseplot
+project_urls =
+    Bug Tracker = https://github.com/toonarmycaptain/parseplot/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: ???
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.9
+install_requires =
+    plusminus>=0.6.0
+
+
+[options.packages.find]
+where = src
